Extract utils from common
