+++ src/__tests__/types/path/common.test-d.ts
@@ -1,9 +1,6 @@
 import { expectType } from 'tsd';
 
 import {
-  AccessPattern,
-  ArrayKey,
-  AsKey,
   AsPathTuple,
   CheckKeyConstraint,
   ContainsIndexable,
@@ -11,9 +8,7 @@ import {
   GetPath,
   HasKey,
   HasPath,
-  IsTuple,
   JoinPathTuple,
-  Key,
   Keys,
   NumericKeys,
   ObjectKeys,
@@ -21,11 +16,10 @@ import {
   SetKey,
   SetPath,
   SplitPathString,
-  ToKey,
   TupleKeys,
-  UnionToIntersection,
   ValidPathPrefix,
 } from '../../../types/path/common';
+import { AccessPattern, Key } from '../../../types/path/internal/utils';
 import {
   _,
   HundredPathString,
@@ -34,18 +28,6 @@ import {
   Nested,
   NullableInfiniteType,
 } from '../__fixtures__';
-
-/** {@link IsTuple} */ {
-  /** it should evaluate to true if it's a tuple */ {
-    const actual = _ as IsTuple<[string, number]>;
-    expectType<true>(actual);
-  }
-
-  /** it should evaluate to false if it's not a tuple */ {
-    const actual = _ as IsTuple<string[]>;
-    expectType<false>(actual);
-  }
-}
 
 /** {@link TupleKeys} */ {
   /** it should evaluate to the own keys of the tuple */ {
@@ -56,35 +38,6 @@ import {
   /** it should evaluate to never if an array type is passed */ {
     const actual = _ as TupleKeys<string[]>;
     expectType<never>(actual);
-  }
-}
-
-/** {@link AsKey} */ {
-  /** it should behave like a noop type when a Key is passed */ {
-    const actual = _ as AsKey<'foo'>;
-    expectType<'foo'>(actual);
-  }
-
-  /** it should evaluate to never if not a Key is passed */ {
-    const actual = _ as AsKey<boolean>;
-    expectType<never>(actual);
-  }
-}
-
-/** {@link ToKey} */ {
-  /** it should behave like a noop type when a Key is passed */ {
-    const actual = _ as ToKey<'foo'>;
-    expectType<'foo'>(actual);
-  }
-
-  /** it should evaluate to never if not a Key or ArrayKey is passed */ {
-    const actual = _ as ToKey<boolean>;
-    expectType<never>(actual);
-  }
-
-  /** it should convert an ArrayKey to a template literal type */ {
-    const actual = _ as ToKey<ArrayKey>;
-    expectType<`${ArrayKey}`>(actual);
   }
 }
 
@@ -174,42 +127,6 @@ import {
   }
 }
 
-/** {@link AccessPattern} */ {
-  type Extends<A, B> = A extends B ? true : false;
-
-  /** it should extend if it's a subtype */ {
-    const actual = _ as Extends<
-      AccessPattern<'abcd', 'abcd'>,
-      AccessPattern<string, 'abcd'>
-    >;
-    expectType<true>(actual);
-  }
-
-  /** it should extend if it's a supertype */ {
-    const actual = _ as Extends<
-      AccessPattern<'abcd', string>,
-      AccessPattern<'abcd', 'abcd'>
-    >;
-    expectType<true>(actual);
-  }
-
-  /** it shouldn't extend if it isn't a subtype */ {
-    const actual = _ as Extends<
-      AccessPattern<string, 'abcd'>,
-      AccessPattern<'abcd', 'abcd'>
-    >;
-    expectType<false>(actual);
-  }
-
-  /** it shouldn't extend if it isn't a supertype */ {
-    const actual = _ as Extends<
-      AccessPattern<'abcd', 'abcd'>,
-      AccessPattern<'abcd', string>
-    >;
-    expectType<false>(actual);
-  }
-}
-
 /** {@link CheckKeyConstraint} */ {
   /** it should remove the keys which don't match the constraint */ {
     const actual = _ as CheckKeyConstraint<
@@ -218,18 +135,6 @@ import {
       AccessPattern<string>
     >;
     expectType<'foo'>(actual);
-  }
-}
-
-/** {@link UnionToIntersection} */ {
-  /** it should intersect a union of objects */ {
-    const actual = _ as UnionToIntersection<{ foo: string } | { bar: number }>;
-    expectType<{ foo: string } & { bar: number }>(actual);
-  }
-
-  /** it should intersect wrapped unions */ {
-    const actual = _ as UnionToIntersection<[0 | 1] | [1 | 2]>[never];
-    expectType<1>(actual);
   }
 }
